2차 리뷰 반영
